# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CategoryRequest:
    properties:
      nama_category:
        type: string
    type: object
...
  models.RegisterRequest:
    properties:
      email:
        type: string
      id_kota:
        type: string
      id_provinsi:
        type: string
      jenis_kelamin:
        type: string
      kata_sandi:
        type: string
      nama:
        type: string
      notelp:
        type: string
      pekerjaan:
        type: string
      tanggal_lahir:
        description: YYYY-MM-DD
        type: string
      tentang:
        type: string
    type: object
  models.UpdateRequest:
    properties:
      nama_category:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
  description: PBI API Documentation
  termsOfService: http://swagger.io/terms/
  title: PBI API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login using email and password
      produces:
      - application/json
      responses:
        "200":
          description: Success login
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Invalid email or password
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create new user account
      parameters:
      - description: Register payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success register user
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "409":
          description: Email already exists
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Register user
      tags:
      - Auth
  /category:
    get:
      description: Get list of categories (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Success get category list
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get all categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create new category (Admin only)
      parameters:
      - description: Category payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success create category
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Category
  /category/{id}:
    delete:
      description: Delete category by ID (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success delete category
          schema:
            type: object
        "400":
          description: Invalid ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Category not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - Category
    get:
      description: Get category detail (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success get category detail
          schema:
            type: object
        "400":
          description: Invalid ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Category not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get category by ID
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update category by ID (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success update category
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Category not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Category
...
  /provcity/detailcity/{province_id}/cities/{city_id}:
    get:
      consumes:
      - application/json
      description: Get detailed information of a specific city
      parameters:
      - description: Province ID
        in: path
        name: province_id
        required: true
        type: string
      - description: City ID
        in: path
        name: city_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success get city detail
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: City not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get City Detail
      tags:
      - Address
  /provcity/detailprovince/{province_id}:
    get:
      consumes:
      - application/json
      description: Get detailed information of a specific province
      parameters:
      - description: Province ID
        in: path
        name: province_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success get province detail
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Province not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Province Detail
      tags:
      - Address
  /provcity/listcities/{province_id}:
    get:
      consumes:
      - application/json
      description: Get list of cities in a specific province
      parameters:
      - description: Province ID
        in: path
        name: province_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success get cities
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Cities by Province
      tags:
      - Address
  /provcity/listprovincies:
    get:
      consumes:
      - application/json
      description: Get list of all provinces in Indonesia
      produces:
      - application/json
      responses:
        "200":
          description: Success get provinces
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get All Provinces
      tags:
      - Address
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"